# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/orders/{orderId}:
    get:
      operationId: OrdersController_getOrder
      parameters: []
      responses:
        '200':
          description: Customer-owned order detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailEnvelopeDto'
      security: *a35
      tags: *a36
...
  /v1/rider/onboard:
    post:
      operationId: RidersController_onboard
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardRiderDto'
      responses:
        '201':
          description: Provisions the rider record for the caller. The client must refresh
            its session afterwards to receive the rider_id claim. The rider
            starts pending and requires admin verification before delivery
            access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderOnboardEnvelopeDto'
        '400':
          description: Invalid onboarding input or unknown campus.
        '401': &a67
          description: Missing, invalid, or expired Supabase JWT.
        '403': &a68
          description: Verified active rider access is required.
        '409':
          description: This account is already linked to a rider.
      security: &a69
        - supabaseAuth: []
      tags: &a70
        - rider
  /v1/rider/profile:
    get:
      operationId: RidersController_getProfile
      parameters: []
      responses:
        '200':
          description: Authenticated rider profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderProfileEnvelopeDto'
        '401': *a67
        '403': *a68
        '404':
          description: Rider profile not found.
      security: *a69
      tags: *a70
    patch:
      operationId: RidersController_updateProfile
      parameters: []
      responses:
        '200':
          description: Updated rider profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderProfileEnvelopeDto'
        '400':
          description: Invalid profile input.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/availability:
    patch:
      operationId: RidersController_setAvailability
      parameters: []
      responses:
        '200':
          description: Updated rider availability flag.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderProfileEnvelopeDto'
        '400':
          description: Invalid availability input.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/payout-account:
    get:
      operationId: RidersController_getPayoutAccount
      parameters: []
      responses:
        '200':
          description: Current masked payout account snapshot, if configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderPayoutAccountEnvelopeDto'
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
    put:
      operationId: RidersController_updatePayoutAccount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRiderPayoutAccountDto'
      responses:
        '200':
          description: Provisions a Paystack transfer recipient from the full account
            number and stores a masked snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderPayoutAccountEnvelopeDto'
        '400':
          description: Invalid payout account input.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/payout-account/verify:
    post:
      operationId: RidersController_verifyPayoutAccount
      parameters: []
      responses:
        '200':
          description: Re-attests the current payout account and refreshes its verification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderPayoutAccountEnvelopeDto'
        '400':
          description: No provisioned payout account to verify.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/payout-history:
    get:
      operationId: RidersController_getPayoutHistory
      parameters: []
      responses:
        '200':
          description: Cursor-paginated rider payout transfer history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderPayoutHistoryEnvelopeDto'
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/assignments:
    get:
      operationId: RidersController_listAssignments
      parameters: []
      responses:
        '200':
          description: Cursor-paginated assignments for the authenticated rider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderAssignmentListEnvelopeDto'
        '400':
          description: Invalid filters or cursor.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/assignments/{assignmentId}:
    get:
      operationId: RidersController_getAssignment
      parameters:
        - name: assignmentId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Assignment detail with order manifest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderAssignmentDetailEnvelopeDto'
        '401': *a67
        '403': *a68
        '404':
          description: Assignment not found for this rider.
      security: *a69
      tags: *a70
  /v1/rider/assignments/{assignmentId}/accept:
    post:
      operationId: RidersController_acceptAssignment
      parameters:
        - name: assignmentId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Assignment accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderAssignmentDetailEnvelopeDto'
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/assignments/{assignmentId}/decline:
    post:
      operationId: RidersController_declineAssignment
      parameters:
        - name: assignmentId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Assignment declined; the batch reopens for admin re-assignment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderAssignmentDetailEnvelopeDto'
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/assignments/{assignmentId}/picked-up:
    post:
      operationId: RidersController_markAssignmentPickedUp
      parameters:
        - name: assignmentId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Assignment marked picked up.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderAssignmentDetailEnvelopeDto'
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/orders/{orderId}:
    get:
      operationId: RidersController_getOrder
      parameters:
        - name: orderId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Delivery order detail for an assigned rider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderOrderDetailEnvelopeDto'
        '401': *a67
        '403': *a68
        '404':
          description: Order not found for this rider.
      security: *a69
      tags: *a70
...
  /v1/rider/orders/{orderId}/issues:
    post:
      operationId: RidersController_createIssue
      parameters:
        - name: orderId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Delivery issue recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderIssueEnvelopeDto'
        '400':
          description: Invalid issue payload.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/earnings:
    get:
      operationId: RidersController_getEarnings
      parameters: []
      responses:
        '200':
          description: Rider earnings grouped by assignment batch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderEarningsEnvelopeDto'
        '400':
          description: Invalid earnings date range.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/settlements:
    get:
      operationId: RidersController_listSettlements
      parameters: []
      responses:
        '200':
          description: Cursor-paginated rider settlement list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderSettlementListEnvelopeDto'
        '400':
          description: Invalid settlement filters or cursor.
        '401': *a67
        '403': *a68
      security: *a69
      tags: *a70
  /v1/rider/settlements/{id}:
    get:
      operationId: RidersController_getSettlement
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Rider settlement detail with settlement lines.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderSettlementDetailEnvelopeDto'
        '401': *a67
        '403': *a68
        '404':
          description: Settlement not found for this rider.
      security: *a69
      tags: *a70
...
  /v1/admin/riders/{riderId}/assignments:
    get:
      operationId: AdminController_getRiderAssignments
      parameters: []
      responses:
        '200':
          description: ''
        '401': *a80
        '403': *a81
      security: *a82
      tags: *a83
  /v1/admin/riders/{riderId}/settlements:
    get:
      operationId: AdminController_getRiderSettlements
      parameters: []
      responses:
        '200':
          description: ''
        '401': *a80
        '403': *a81
      security: *a82
      tags: *a83
...
info:
  title: Meal Direct API
  description: Stable REST API for Meal Direct web and future mobile clients.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    supabaseAuth:
      scheme: bearer
      bearerFormat: Supabase JWT
      type: http
      description: Supabase Auth access token.
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Required for important mutations such as order creation and payments.
    requestId:
      type: apiKey
      in: header
      name: x-request-id
      description: Optional caller-provided request ID. The API echoes or generates one.
    operationsToken:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Internal operations token. Temporary until Super Admin JWT/RBAC is
        implemented.
  schemas:
...
    RiderProfileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campusId:
          type: string
          format: uuid
        displayName:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
            - pending
            - verified
            - suspended
            - deactivated
        active:
          type: boolean
        available:
          type: boolean
      required:
        - id
        - campusId
        - displayName
        - phone
        - status
        - active
        - available
...
    OrderDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        customerId:
          type: string
          format: uuid
        customerDisplayName:
          type: string
          nullable: true
          description: Customer display name. Present only on vendor/admin/rider order
            views.
        customerPhone:
          type: string
          nullable: true
          description: Customer phone number. Present only on vendor/admin/rider order
            views.
        campusId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        vendorDisplayName:
          type: string
        serviceDate:
          type: string
        deliverySlotId:
          type: string
          format: uuid
        deliverySlotName:
          type: string
        locationId:
          type: string
          format: uuid
        locationName:
          type: string
        orderStatus:
          type: string
          enum: *a97
        deliveryMode:
          type: string
        specialInstructions:
          type: string
          nullable: true
        roomNumber:
          type: string
          nullable: true
          description: Hostel room number when the delivery location is a hostel.
        foodSubtotalKobo:
          type: number
        deliveryFeeKobo:
          type: number
        serviceFeeKobo:
          type: number
        discountKobo:
          type: number
        largeOrderSurchargeKobo:
          type: number
          description: Large-order surcharge (1.5% + ₦100) charged on this order. Included
            in totalKobo.
        totalKobo:
          type: number
        currency:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        paidAt:
          type: string
          nullable: true
        deliveredAt:
          type: string
          nullable: true
        confirmedAt:
          type: string
          nullable: true
        deliveryCode:
          type: string
          nullable: true
          description: Delivery hand-off code to read to the rider after receiving the
            order.
        deliveryHandoff:
          nullable: true
          description: Customer-facing code + instruction payload for delivery hand-off.
          type: object
          allOf:
            - $ref: '#/components/schemas/DeliveryHandoffDto'
        batchId:
          type: string
          nullable: true
          format: uuid
          description: Delivery batch this order belongs to, when batched. Powers the
            batch chat.
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDto'
        latestPayment:
          nullable: true
          type: object
          allOf:
            - $ref: '#/components/schemas/PaymentSnapshotDto'
      required:
        - id
        - orderNumber
        - customerId
        - campusId
        - vendorId
        - vendorDisplayName
        - serviceDate
        - deliverySlotId
        - deliverySlotName
        - locationId
        - locationName
        - orderStatus
        - deliveryMode
        - foodSubtotalKobo
        - deliveryFeeKobo
        - serviceFeeKobo
        - discountKobo
        - largeOrderSurchargeKobo
        - totalKobo
        - currency
        - createdAt
        - updatedAt
        - items
...
    OrderDetailEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/OrderDetailDto'
      required:
        - data
...
    OnboardRiderDto:
      type: object
      properties:
        campusId:
          type: string
          format: uuid
          description: Campus the rider delivers on.
        displayName:
          type: string
          maxLength: 120
          minLength: 2
          description: Public display name.
        phone:
          type: string
          pattern: ^[+0-9][0-9 ()-]{6,24}$
      required:
        - campusId
        - displayName
        - phone
    RiderOnboardResultDto:
      type: object
      properties:
        rider:
          $ref: '#/components/schemas/RiderProfileDto'
        tokenRefreshRequired:
          type: boolean
          description: The client must refresh its session to receive the rider_id claim.
      required:
        - rider
        - tokenRefreshRequired
    RiderOnboardEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RiderOnboardResultDto'
      required:
        - data
    RiderProfileEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RiderProfileDto'
      required:
        - data
    RiderPayoutAccountDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        riderId:
          type: string
          format: uuid
        paystackRecipientCode:
          type: string
          nullable: true
        bankName:
          type: string
        bankCode:
          type: string
          nullable: true
        maskedAccountNumber:
          type: string
        accountName:
          type: string
        verifiedAt:
          type: string
          nullable: true
        adminReviewStatus:
          type: string
          enum:
            - pending
            - approved
            - rejected
        failureReason:
          type: string
          nullable: true
        verificationStatus:
          type: string
          enum:
            - verified
            - unverified
        payoutMode:
          type: string
          enum:
            - manual
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - riderId
        - bankName
        - maskedAccountNumber
        - accountName
        - adminReviewStatus
        - verificationStatus
        - payoutMode
        - active
        - createdAt
        - updatedAt
    RiderPayoutAccountEnvelopeDto:
      type: object
      properties:
        data:
          nullable: true
          type: object
          allOf:
            - $ref: '#/components/schemas/RiderPayoutAccountDto'
      required:
        - data
    UpsertRiderPayoutAccountDto:
      type: object
      properties:
        bankName:
          type: string
          maxLength: 120
          minLength: 1
        bankCode:
          type: string
          maxLength: 20
          minLength: 1
          description: Paystack bank code, required to provision the transfer recipient.
        accountName:
          type: string
          maxLength: 160
          minLength: 1
        accountNumber:
          type: string
          description: Full account number used to provision the recipient; only a mask is
            stored.
      required:
        - bankName
        - bankCode
        - accountName
        - accountNumber
    RiderPayoutTransferDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        settlementId:
          type: string
          format: uuid
        settlementDate:
          type: string
          nullable: true
        reference:
          type: string
        amountKobo:
          type: number
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - settlementId
        - reference
        - amountKobo
        - status
        - createdAt
        - updatedAt
    RiderPayoutHistoryEnvelopeDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RiderPayoutTransferDto'
      required:
        - data
...
    RiderAssignmentListEnvelopeDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RiderAssignmentSummaryDto'
      required:
        - data
...
    RiderAssignmentDetailEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RiderAssignmentDetailDto'
      required:
        - data
...
    RiderOrderDetailEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RiderOrderDetailDto'
      required:
        - data
...
    RiderIssueEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RiderIssueDto'
      required:
        - data
...
    RiderEarningsEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RiderEarningsSummaryDto'
      required:
        - data
...
    RiderSettlementListEnvelopeDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RiderSettlementSummaryDto'
      required:
        - data
...
    RiderSettlementDetailEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RiderSettlementDetailDto'
      required:
        - data
...